% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interactive.R
\name{check_interactive}
\alias{check_interactive}
\title{Verify Correctness of User-Specified Interactive Preference}
\usage{
check_interactive(interactive)
}
\arguments{
\item{interactive}{Logical. Should program be run interactively?}
}
\description{
Verify Correctness of User-Specified Interactive Preference
}
\keyword{internal}
