% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{comp_lookup}
\alias{comp_lookup}
\title{Helper function for intensity coding process.}
\usage{
comp_lookup(incorrect_entries, Levels, compendium)
}
\arguments{
\item{incorrect_entries}{A vector of activities that have not been correctly coded yet}

\item{Levels}{A vector of intensity levels from which to select}

\item{compendium}{A version compendium, passed from
\code{compendium_reference}, that has been modified to reflect the
intensity selections made in that function}
}
\description{
Interface for looking up Compendium values to assign an intensity to an
activity.
}
\keyword{internal}
