% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direct Observation Program and Functions.R
\name{compendium_reference}
\alias{compendium_reference}
\title{Consult
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities}
\usage{
compendium_reference(data)
}
\arguments{
\item{data}{A data frame outputted from \code{\link{data_collection_program}}}
}
\value{
A data frame fully annotated with intensity values
}
\description{
Consult
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities
}
\examples{
\dontrun{
observation_data <- data_collection_program()
compendium_reference(observation_data)
}
}
