% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Manual.R
\name{compendium_reference_manual}
\alias{compendium_reference_manual}
\title{Manually Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities}
\usage{
compendium_reference_manual(obs_data)
}
\arguments{
\item{obs_data}{A data frame outputted from
\code{\link{data_collection_program_manual}}}
}
\value{
A data frame fully annotated with intensity values
}
\description{
Manually Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities
}
\keyword{internal}
