% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{data_collection_program}
\alias{data_collection_program}
\title{Interactively Collect Direct Observation Data}
\usage{
data_collection_program(...)
}
\arguments{
\item{...}{Arguments passed to dialog functions}
}
\value{
A data frame of observation data
}
\description{
Interactively Collect Direct Observation Data
}
\note{
This program relies heavily on dialog boxes generated by the
  `svDialogs` package. Default measures are in place to ensure baseline
  functionality across operating systems and R interfaces. However, a better
  experience may be achievable by passing additional arguments to the dialog
  functions via `...`, as the vignette for `Observation` describes.
}
\examples{
\dontrun{
data_collection_program()
}

}
\seealso{
[example_data]
}
