% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{data_collection_program_interactive}
\alias{data_collection_program_interactive}
\title{Interactively Collect Direct Observation Data}
\usage{
data_collection_program_interactive(...)
}
\arguments{
\item{...}{Arguments passed to dialog functions}
}
\value{
A data frame of observation data
}
\description{
Interactively Collect Direct Observation Data
}
\note{
\code{Observation:::data_collection_program_interactive()} is
    called within \code{data_collection_program()} when
    \code{interactive = TRUE} and the session is interactive.

    This program relies heavily on dialog boxes generated by the
    `svDialogs` package. Default measures are in place to ensure
    baseline
    functionality across operating systems and R interfaces.
    However, a better experience may be achievable by passing
    additional arguments to the dialog functions via `...`, as the
    vignette for `Observation` describes.
}
\examples{
if (interactive()) {
data_collection_program()
}

}
\seealso{
[example_data]

Other collection functions: \code{\link{data_collection_program_manual}},
  \code{\link{data_collection_program}}
}
\keyword{internal}
