% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Manual.R
\name{data_collection_program_manual}
\alias{data_collection_program_manual}
\title{Manually Collect Direct Observation Data}
\usage{
data_collection_program_manual()
}
\description{
Manually Collect Direct Observation Data
}
\examples{
Observation:::data_collection_program_manual()
}
\keyword{internal}
