% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Sample data collected interactively with direct observation program.}
\format{A data frame with 10 rows and 15 variables:
\describe{
  \item{id}{The user-inputted identification code for the trial}
  \item{User_Timestamp}{The user-inputted start time for the activity}
  \item{Auto_Timestamp}{The automatic timestamp generated by button click, retained as a backup if an error occurs in User_Timestamp}
  \item{dayofyear}{The numeric day of the year (i.e., julian date)}
  \item{minofday}{The numeric minute of the day (e.g. 0 for midnight and 1439 for 23:59)}
  \item{Activity}{The user-inputted activity description}
  \item{duration_s}{The duration of the activity (in seconds), based on User_Timestamp}
  \item{auto_duration_s}{The duration of the activity (in seconds), based on Auto_Timestamp}
  \item{Tree_Intensity}{The pre-classified activity intensity, based on user input to the program's decision tree}
  \item{seated}{User-inputted response to the seated prompt}
  \item{large_muscles_moving}{User-inputted response to the large-muscles-moving prompt}
  \item{slow}{User-inputted response to the slow prompt}
  \item{slowed_by_resistance}{User-inputted response to the slowed-by-resistance prompt}
  \item{ambulation}{User-inputted response to the ambulation prompt}
  \item{light_walking}{User-inputted response to the light-walking prompt}
}}
\usage{
data(example_data, package = "Observation")
}
\description{
A data set assembled by running the interactive direct observation program
and entering a variety of activities.
}
\keyword{datasets}
