% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{interactive_compendium_reference}
\alias{interactive_compendium_reference}
\title{Interactively Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities}
\usage{
interactive_compendium_reference(obs_data, ...)
}
\arguments{
\item{obs_data}{A data frame outputted from
\code{\link{interactive_data_collection_program}}}

\item{...}{Arguments passed to dialog functions}
}
\value{
A data frame fully annotated with intensity values
}
\description{
Interactively Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities
}
\note{
\code{interactive_compendium_reference} is called
    within \code{\link{compendium_reference}} when
    \code{interactive = TRUE} and the session is interactive.
}
\examples{

if (interactive()) {
data(example_data)
compendium_reference(example_data)

observation_data <- data_collection_program()
compendium_reference(observation_data)
}
}
\seealso{
Other processing functions: 
\code{\link{compendium_reference}()},
\code{\link{manual_compendium_reference}()}
}
\concept{processing functions}
\keyword{internal}
