% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{interactive_tree_intensity}
\alias{interactive_tree_intensity}
\title{Interactively Pre-Classify Activity Intensity}
\usage{
interactive_tree_intensity(...)
}
\arguments{
\item{...}{Arguments passed to dialog functions}
}
\description{
Interactively implement the pre-classification decision tree
    described at the end of Supplemental Document 3 from
\href{https://pubmed.ncbi.nlm.nih.gov/29135657/}{Hibbing et al. (2018,
*Med Sci Sports Exerc*)}.
}
\keyword{internal}
