% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direct Observation Program and Functions.R
\name{tree_intensity}
\alias{tree_intensity}
\title{Pre-classify activity intensity}
\usage{
tree_intensity()
}
\description{
Pre-classify activity intensity
}
\keyword{internal}
