% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{tree_intensity_interactive}
\alias{tree_intensity_interactive}
\title{Interactively Pre-Classify Activity Intensity}
\usage{
tree_intensity_interactive(...)
}
\arguments{
\item{...}{Arguments passed to dialog functions}
}
\description{
Interactively implement the pre-classification decision tree described at the
end of Supplemental Document 3 from
\href{https://www.ncbi.nlm.nih.gov/pubmed/29135657}{Hibbing et al. (2018,
*Med Sci Sports Exerc*)}.
}
\keyword{internal}
