% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Manual.R
\name{tree_intensity_manual}
\alias{tree_intensity_manual}
\title{Manually Pre-Classify Activity Intensity}
\usage{
tree_intensity_manual()
}
\description{
Manually implement the pre-classification decision tree described at the end
of Supplemental Document 3 from
\href{https://www.ncbi.nlm.nih.gov/pubmed/29135657}{Hibbing et al. (2018,
*Med Sci Sports Exerc*)}.
}
\examples{
tree_intensity_manual
}
\keyword{internal}
