% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{compendium_reference}
\alias{compendium_reference}
\title{Consult
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Interactively Encode Direct Observation Intensities}
\usage{
compendium_reference(obs_data)
}
\arguments{
\item{obs_data}{A data frame outputted from
\code{\link{data_collection_program}}}
}
\value{
A data frame fully annotated with intensity values
}
\description{
Consult
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Interactively Encode Direct Observation Intensities
}
\examples{

\dontrun{
data(example_data)
compendium_reference(example_data)

observation_data <- data_collection_program()
compendium_reference(observation_data)
}
}
