% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{compendium_reference_interactive}
\alias{compendium_reference_interactive}
\title{Interactively Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities}
\usage{
compendium_reference_interactive(obs_data, ...)
}
\arguments{
\item{obs_data}{A data frame outputted from
\code{\link{data_collection_program_interactive}}}

\item{...}{Arguments passed to dialog functions}
}
\value{
A data frame fully annotated with intensity values
}
\description{
Interactively Consult the
\href{https://sites.google.com/site/compendiumofphysicalactivities/}{Compendium
of Physical Activities} to Encode Direct Observation Intensities
}
\examples{

if (interactive()) {
data(example_data)
compendium_reference_interactive(example_data)

observation_data <- data_collection_program_interactive()
compendium_reference_interactive(observation_data)
}
}
\keyword{internal}
