% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{data_collection_program}
\alias{data_collection_program}
\title{Interactively Collect Direct Observation Data}
\usage{
data_collection_program()
}
\value{
A data frame of observation data
}
\description{
Interactively Collect Direct Observation Data
}
\examples{
\dontrun{
data_collection_program()
}

}
\seealso{
[example_data]
}
