% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudo_Generics.R
\name{data_collection_program}
\alias{data_collection_program}
\title{Collect Direct Observation Data}
\usage{
data_collection_program(interactive = TRUE, ...)
}
\arguments{
\item{interactive}{Logical. Should program be run interactively?}

\item{...}{Additional arguments passed to method-like functions (under "See Also" below)}
}
\description{
A generic-like function for collecting direct observation data manually or
interactively, depending on user-specified preference and session
capabilities.
}
\note{
If `interactive = TRUE`, but R is not being used interactively, a
  message is issued and an override is implemented to set `interactive =
  FALSE`.
}
\examples{
if (interactive()) {
data_collection_program()
}

data_collection_program(interactive = FALSE)
}
\seealso{
[data_collection_program_interactive()],
    [data_collection_program_manual()]
}
