% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direct Observation Program and Functions.R
\name{data_collection_program}
\alias{data_collection_program}
\title{Collect direct observation data}
\usage{
data_collection_program()
}
\value{
A data frame of observation data
}
\description{
Collect direct observation data
}
\examples{
\dontrun{
data_collection_program()
}
}
