% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Program_&Functions.R
\name{tree_intensity}
\alias{tree_intensity}
\title{Pre-classify activity intensity}
\usage{
tree_intensity()
}
\description{
This is the decision tree to pre-classify intensity.
See the Figure at the end of Supplemental Document 3 from
\href{https://www.ncbi.nlm.nih.gov/pubmed/29135657}{Hibbing et al.
(2018, *Med Sci Sports Exerc*)}.
}
\keyword{internal}
